refactor(d11): split settings.rs into module dir (parse + io)
Review finding: post-Task-3, settings.rs sat at ~637 lines — past the
500-line proactive-split threshold the project uses for new code. The
file housed two unrelated concerns under one roof: the Task-2 pure
parse/serialize layer and the Task-3 vault I/O facade + per-vault
device-UUID persistence.

Split into a `settings/` directory:

- `mod.rs` (27 lines): module declarations + a flat re-export surface
  so external callers (`session.rs`, integration tests) keep using
  `crate::settings::Settings` etc. unchanged. Pure organisational layer.
- `parse.rs` (273 lines): pure value types — `Settings`, `ParseResult`,
  `parse_settings_field`, `serialize_settings`, `validate_save_value`,
  the private `clamp_with_warning` helper, and the existing parse-layer
  test module.
- `io.rs` (477 lines): vault-touching code — `load_from_vault`,
  `save_to_vault`, `find_settings_block_uuid`, the device-UUID
  persistence (`device_uuid_path_in`, `load_or_create_device_uuid`,
  `load_or_create_device_uuid_in`, `read_device_uuid_file`), and the
  I/O-layer test module. Imports `Settings` etc. from `super::parse`.

Test module in `io.rs` renamed from `io_tests` to `tests` to match the
conventional one-`mod tests`-per-file pattern post-split. No spec or
handoff docs reference the old names (greped).

Diff is dominated by file moves; logic is byte-identical to the prior
state. All 38 lib tests + 13 integration tests continue to pass.
